Run unstable tests in separate environment
*Issue #, if available:*764

*Description of changes:*
This commit splits up the troublesome tests of models into two parts, one stable and one unstable part. 
The unstable part of the tests are moved to a new file and are run during a seperate test execution.

The unstable parts of the tests are moved to the folder:

`/tests/model/unstable_tests/test_name.py`

for example the troublesome parts of the file:

`tests/model/deepstate/test_model.py`

is moved to:

`/tests/model/unstable_tests/test_deepstate.py`

In order to run these tests after a pull request ive modified the file:

`.github/workflows/tests.yml`

More specifically the change made should allow for the tests to run in parallel. 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.
