+++ .github/workflows/tests.yml
@@ -11,7 +11,7 @@ jobs:
       matrix:
         python-version: [3.6, 3.7]
         platform: [ubuntu-latest, windows-latest]
-    
+        pytest-parameters: [--ignore=test/model/unstable_tests, test/model/unstable_tests]
     runs-on: ${{ matrix.platform }}
 
     steps:
@@ -30,4 +30,4 @@ jobs:
         pip install -r requirements/requirements-extras-m-competitions.txt
     - name: Test with pytest
       run: |
-        pytest -m 'not (gpu or serial)' --cov src/gluonts --cov-report=term --cov-report xml test
+        pytest -m 'not (gpu or serial)' ${{ matrix.pytest-parameters }} --cov src/gluonts --cov-report=term --cov-report xml test
